@naoh87 2022-02-02
I change warm up time to 60s and run 30s for stabilize result because #486 and #493 performance is so close.

This benchmark was run on MacBookPro 2019 (4 core 2.8GHz Core i7)

This PR
```
Summary:
  Count:	357669
  Total:	30.00 s
  Slowest:	26.08 ms
  Fastest:	0.14 ms
  Average:	2.15 ms
  Requests/sec:	11921.77

Response time histogram:
  0.142  [1]      |
  2.736  [281690] |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  5.330  [63114]  |∎∎∎∎∎∎∎∎∎
  7.924  [12309]  |∎∎
  10.518 [467]    |
  13.112 [41]     |
  15.706 [6]      |
  18.300 [10]     |
  20.894 [0]      |
  23.488 [0]      |
  26.083 [1]      |

Latency distribution:
  10 % in 0.92 ms
  25 % in 1.32 ms
  50 % in 1.84 ms
  75 % in 2.54 ms
  90 % in 3.98 ms
  95 % in 4.94 ms
  99 % in 6.49 ms
```

#486 08f556f

```
Summary:
  Count:	332880
  Total:	30.00 s
  Slowest:	19.52 ms
  Fastest:	0.16 ms
  Average:	2.30 ms
  Requests/sec:	11095.20

Response time histogram:
  0.158  [1]      |
  2.095  [184726] |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  4.031  [110812] |∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎∎
  5.968  [27372]  |∎∎∎∎∎∎
  7.905  [9036]   |∎∎
  9.841  [828]    |
  11.778 [35]     |
  13.715 [13]     |
  15.651 [4]      |
  17.588 [1]      |
  19.525 [2]      |

Latency distribution:
  10 % in 0.97 ms
  25 % in 1.39 ms
  50 % in 1.96 ms
  75 % in 2.72 ms
  90 % in 4.24 ms
  95 % in 5.32 ms
  99 % in 6.93 ms
```