@ahjohannessen 2022-02-02
I see, not much of a difference between those two.

How do you run this `ghz --cpus=3 -n 100000` Using docker?