@naoh87 2022-02-02
Nope. I just `sbt run` because I am not interested in Docker Desktop VM overhead.