@ahjohannessen 2022-02-02
Perhaps using constructs from https://github.com/davenverse/condemned/tree/main/core/src/main/scala/io/chrisdavenport/condemned
can make my take on it a bit faster.