@ahjohannessen 2022-02-02
>Nope. I just sbt run because I am not interested in Docker Desktop VM overhead.

I mean, is ghz something you installed with brew?