@ahjohannessen 2022-02-02
Would you mind sharing the ghz command, so I could try locally here? :)