@naoh87 2022-02-02
`❯ ghz --cpus=3 --proto ../scenarios/complex_proto/helloworld.proto -D ../scenarios/complex_proto/payload --insecure localhost:50051 --call helloworld.Greeter.SayHello -z 30s` at `grpc_bench/scala_fs2_bench`