@naoh87 2022-02-02
I am really sorry about that my above benchmark has completely mistook measurement.
I was forgot to switch ScalaPB implementation.