@naoh87 2022-02-02
I retry benchmark. But both implement are almost same peformance.
This PR
```
Summary:
  Count:	660755
  Total:	60.00 s
  Slowest:	25.08 ms
  Fastest:	0.17 ms
  Average:	2.32 ms
  Requests/sec:	11011.78

Latency distribution:
  10 % in 0.99 ms
  25 % in 1.42 ms
  50 % in 1.98 ms
  75 % in 2.75 ms
  90 % in 4.27 ms
  95 % in 5.31 ms
  99 % in 6.92 ms
```

#486
```
Summary:
  Count:	640781
  Total:	60.01 s
  Slowest:	19.69 ms
  Fastest:	0.19 ms
  Average:	2.44 ms
  Requests/sec:	10678.68

Latency distribution:
  10 % in 1.07 ms
  25 % in 1.53 ms
  50 % in 2.12 ms
  75 % in 2.90 ms
  90 % in 4.39 ms
  95 % in 5.46 ms
  99 % in 7.11 ms
```