@naoh87 2022-02-02
I was also tried start `run(message)` at `onMessage` and give up handle `Too many requests` but it does not improve performance.