@naoh87 2022-02-02
This result is another run one, but it's enough to use.
```
Status code distribution:
  [OK]            640219 responses
  [Canceled]      12 responses
  [Unavailable]   36 responses
```