@naoh87 2022-02-03
I benchmarked with `-XX:+UnlockDiagnosticVMOptions` and `-XX:+DebugNonSafepoints` and profile with async-profiler. 
This shows #493 is less runtime overhead.

#493
```
Summary:
  Count:	675933
  Total:	60.01 s
  Slowest:	24.32 ms
  Fastest:	0.17 ms
  Average:	2.29 ms
  Requests/sec:	11264.55

Latency distribution:
  10 % in 0.98 ms
  25 % in 1.41 ms
  50 % in 1.96 ms
  75 % in 2.70 ms
  90 % in 4.22 ms
  95 % in 5.24 ms
  99 % in 6.79 ms
```

#486
```
Summary:
  Count:	639397
  Total:	60.00 s
  Slowest:	103.10 ms
  Fastest:	0.19 ms
  Average:	2.44 ms
  Requests/sec:	10655.82

Latency distribution:
  10 % in 1.06 ms
  25 % in 1.52 ms
  50 % in 2.11 ms
  75 % in 2.88 ms
  90 % in 4.41 ms
  95 % in 5.49 ms
  99 % in 7.10 ms
```